--- src/utils/ingredientsApiRequest.js
@@ -1,9 +1,0 @@
-export const ingredientsApiRequest = (setState, setError) => {
-  const url = 'https://norma.nomoreparties.space/api/ingredients'
-  fetch(url)
-    .then((response) =>
-      response.ok ? response.json() : Promise.reject(response)
-    )
-    .then((response) => setState(response.data))
-    .catch((error) => setError(error))
-}
